requirements-api.txt
kafka-python
psycopg2
google-api-python-client
requests
fuzzywuzzy
bs4
responses
flask
flask_cors